Abbreviate 'minutes' to 'min' in Moment.
Leverage `moment.updateLocale` to abbreviate 'minutes' to 'min' as rendered in WaitDetails.jsx. This change makes it possible to fit the estimated wait in one row of text in a mobile view. The configuration is placed in index.js since Moment requires this change to be global configuration, and the function does not need to be called with each render cycle.